+++ pkg/conf/parse/parse.go
@@ -65,7 +65,6 @@ var requireRestart = []string{
 	"auth.disableAccessTokens",
 	"auth.providers",
 	"externalURL",
-	"tls.letsencrypt",
 	"git.cloneURLToRepositoryName",
 	"searchScopes",
 	"extensions",